+++ pkg/util/constants/constants.go
@@ -58,6 +58,11 @@ const (
 	FinalizerPVCMemberExists           = "pvc.database.arangodb.com/member-exists"       // Finalizer added to PVCs, indicating the need to keep is as long as its member exists
 	FinalizerDelayPodTermination       = "pod.database.arangodb.com/delay"               // Finalizer added to Pod, delays termination
 
+	AnnotationShutdownManagedContainer      = "shutdown.arangodb.com/managed"
+	AnnotationShutdownContainer             = "container.shutdown.arangodb.com"
+	AnnotationShutdownCoreContainer         = "core.shutdown.arangodb.com"
+	AnnotationShutdownCoreContainerModeWait = "wait"
+
 	AnnotationEnforceAntiAffinity = "database.arangodb.com/enforce-anti-affinity" // Key of annotation added to PVC. Value is a boolean "true" or "false"
 
 	BackupLabelRole      = "backup/role"